src/requirements.txt
requests==2.32.5
python-dotenv==1.1.1
openai==2.3.0
pandas==2.3.3
azure-cosmos==4.9.0
semantic-kernel==1.37.0
azure-ai-projects==1.1.0b4
azure-ai-agents==1.2.0b5
azure-identity ==1.25.1
pyodbc==5.2.0
azure-kusto-data==5.0.5
opentelemetry-sdk==1.37.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-monitor-opentelemetry==1.8.1
opentelemetry-instrumentation-openai-v2==2.1b0
Pillow==11.3.0
sqlalchemy==2.0.44
opencv-python==4.12.0.88
azure-search-documents==11.6.0
fastapi==0.119.0
uvicorn[standard]==0.37.0
azure.ai.inference==1.0.0b9
orjson==3.11.3
numpy==2.2.6
azure-ai-evaluation[redteam]==1.12.0
a2a-sdk[sqlite]==0.3.8
sse-starlette==3.0.2
starlette==0.48.0